Optimize insertion of items into BVH: flat sort of hashes
Previously we were first partitioning the hashes into levels and then
sorting each level. But sorting by node keys induces sorting by levels.
Therefore we can more efficiently sort and partition by first sorting
the whole array of hashes and then doing bisection searches for the
level partitions.

Also remove adaptation of bounding box of BVH since this was not
implemented properly. If we want to do this we have to do it right.